# docs/swagger.yaml
basePath: /
definitions:
...
  admin.SavePlanRequest:
    properties:
      cycle:
        type: string
      description:
        type: string
      features:
        items:
          type: string
        type: array
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      storage_limit:
        type: integer
      upload_limit:
        type: integer
    required:
    - cycle
    - name
    - price
    - storage_limit
    - upload_limit
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for Stream application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Stream API
  version: "1.0"
paths:
...
  /admin/plans:
    get:
      description: Get all plans with usage counts (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List Plans
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Create a plan (admin only)
      parameters:
      - description: Plan payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.SavePlanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create Plan
      tags:
      - admin
  /admin/plans/{id}:
    delete:
      description: Delete a plan, or deactivate it if already used (admin only)
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete Plan
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Update a plan (admin only)
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: Plan payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.SavePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update Plan
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"